+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ###Added
 - Add `resource_group` optional field for metrics
+- Support for dedicated virtual machine hosts
 
 ## 3.39.0 (August 21, 2019)